+++ mowl/datasets/el/__init__.py
@@ -1,3 +1,5 @@
-from .el_dataset import ELDataset, GCI0Dataset, GCI1Dataset, GCI2Dataset, GCI3Dataset
+from .el_dataset import ELDataset, GCI0Dataset, GCI1Dataset, GCI2Dataset, GCI3Dataset, \
+    RoleAxiomDataset, RoleInclusionDataset, RoleChainDataset
 
-__all__ = ["ELDataset", "GCI0Dataset", "GCI1Dataset", "GCI2Dataset", "GCI3Dataset"]
+__all__ = ["ELDataset", "GCI0Dataset", "GCI1Dataset", "GCI2Dataset", "GCI3Dataset",
+           "RoleAxiomDataset", "RoleInclusionDataset", "RoleChainDataset"]